feat(signal-overlay): add SignalOverlay component for displaying environmental alerts and data feat(openstreet-map): integrate SignalOverlay into OpenStreetMap component feat(use-responsive): implement responsive hooks for mobile and tablet detection
