+++ src/components/map/openstreet-map-helpers.tsx
@@ -1,98 +1,4 @@
-"use client";
-
-import * as React from "react";
-import { createPortal } from "react-dom";
-import {
-  Menu,
-  Plus,
-  Minus,
-  Locate,
-  Activity,
-  Cloud,
-  CloudLightning,
-  Flame,
-  Maximize,
-  Sun,
-  Moon,
-  Loader2,
-  SlidersHorizontal,
-  Globe2,
-  AlertTriangle,
-  CircleDot,
-  Clock,
-  ExternalLink,
-  Gauge,
-  Map,
-  MapPin,
-  Radio,
-  Signal,
-  Snowflake,
-  Users,
-  Waves,
-  X,
-  Wind,
-  CalendarClock,
-  MapPinned,
-  ScanLine,
-  FlaskConical,
-  Building2,
-  Ruler,
-} from "lucide-react";
-import {
-  useMap,
-} from "@/components/ui/map";
-import { useTheme } from "next-themes";
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from "@/components/ui/tooltip";
-import { Drawer, DrawerContent } from "@/components/ui/drawer";
-import { cn } from "@/lib/utils";
-import { toast } from "sonner";
-import {
-  getLocationErrorMessage,
-  requestCurrentCoordinates,
-} from "@/lib/browser-geolocation";
-import { EarthquakeItem } from "@/components/hazr-earthquake-item";
-import { HourlyForecastDock } from "@/components/map/hourly-forecast-dock";
-import { WeatherDock } from "@/components/map/weather-dock";
-import { MobileBottomNav } from "@/components/mobile-bottom-nav";
-import { GlobalActivity } from "@/components/global-activity";
-import { HazrSettingsPanel } from "@/components/hazr-settings-panel";
-import { HazrAboutContent } from "@/components/hazr-about-dialog";
-import type {
-  ProcessedAirQualitySite,
-  ProcessedEarthquake,
-  ProcessedEonetEvent,
-  ProcessedTsunamiAlert,
-} from "@/types/api";
-import { getMagnitudeColor, getMagnitudeLabel } from "@/types/api";
-import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
-import {
-  DropdownMenu,
-  DropdownMenuCheckboxItem,
-  DropdownMenuContent,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import {
-  Pagination,
-  PaginationButton,
-  PaginationContent,
-  PaginationEllipsis,
-  PaginationItem,
-  PaginationNext,
-  PaginationPrevious,
-} from "@/components/ui/pagination";
-import type { AppSettings, DataLayerVisibility } from "@/types/settings";
-import { useMap3DMode } from "@/hooks/use-map-3d-mode";
-import { MapLayersControl } from "@/components/map/controls/map-layers-control";
-import { MapCameraDock } from "@/components/map/controls/map-camera-dock";
-
+import type { DataLayerVisibility } from "@/types/settings";
 
 export type MapViewState = {
   center: [number, number];
@@ -101,72 +7,7 @@ export type MapViewState = {
 
 export type LayerVisibility = DataLayerVisibility;
 
-export const useIsMobile = () => {
-  const [isMobile, setIsMobile] = React.useState(false);
-
-  React.useEffect(() => {
-    const media = window.matchMedia("(max-width: 768px)");
-    const handleChange = () => setIsMobile(media.matches);
-    handleChange();
-    media.addEventListener("change", handleChange);
-    return () => media.removeEventListener("change", handleChange);
-  }, []);
-
-  return isMobile;
-};
-
-export const useIsTablet = () => {
-  const [isTablet, setIsTablet] = React.useState(false);
-
-  React.useEffect(() => {
-    const media = window.matchMedia("(max-width: 1024px)");
-    const handleChange = () => setIsTablet(media.matches);
-    handleChange();
-    media.addEventListener("change", handleChange);
-    return () => media.removeEventListener("change", handleChange);
-  }, []);
-
-  return isTablet;
-};
-
-export const getEventIcon = (category: string) => {
-  const normalized = category.toLowerCase();
-  if (normalized.includes("storm")) return CloudLightning;
-  if (normalized.includes("wildfire") || normalized.includes("fire")) return Flame;
-  if (normalized.includes("flood")) return Waves;
-  if (normalized.includes("volcano")) return AlertTriangle;
-  if (normalized.includes("ice")) return Snowflake;
-  if (normalized.includes("drought")) return Sun;
-  if (normalized.includes("dust")) return Wind;
-  return Globe2;
-};
-
-const getEventCategoryLabel = (category: string) => {
-  const normalized = category.toLowerCase();
-  if (normalized.includes("storm")) return "Storm";
-  if (normalized.includes("wildfire") || normalized.includes("fire")) return "Wildfire";
-  if (normalized.includes("flood")) return "Flood";
-  if (normalized.includes("drought")) return "Drought";
-  if (normalized.includes("ice")) return "Iceberg";
-  if (normalized.includes("dust")) return "Dust";
-  if (normalized.includes("volcano")) return "Volcano";
-  return category;
-};
-
-const formatAirQualityParameterLabel = (parameter: string) => {
-  const normalized = parameter.trim().toLowerCase();
-  if (normalized === "pm2.5") return "Fine particles (PM2.5)";
-  if (normalized === "pm10") return "Coarse particles (PM10)";
-  if (normalized === "no2") return "Nitrogen dioxide (NO₂)";
-  if (normalized === "o3") return "Ozone (O₃)";
-  return parameter.toUpperCase();
-};
-
-const formatAirQualityConcentration = (value: number, unit: string) => {
-  return `${value.toFixed(1)} ${unit}`;
-};
-
-const getPaginationRange = (totalPages: number, currentPage: number) => {
+export const getPaginationRange = (totalPages: number, currentPage: number) => {
   if (totalPages <= 5) {
     return Array.from({ length: totalPages }, (_, index) => index + 1);
   }
@@ -181,1519 +22,3 @@ const getPaginationRange = (totalPages: number, currentPage: number) => {
 
   return [1, "ellipsis", currentPage, "ellipsis", totalPages] as const;
 };
-
-type EventToneKey =
-  | "wildfire"
-  | "storm"
-  | "flood"
-  | "volcano"
-  | "ice"
-  | "drought"
-  | "dust"
-  | "default";
-
-const getEventToneKey = (category: string): EventToneKey => {
-  const normalized = category.toLowerCase();
-  if (normalized.includes("wildfire") || normalized.includes("fire")) return "wildfire";
-  if (normalized.includes("storm")) return "storm";
-  if (normalized.includes("flood") || normalized.includes("tsunami")) return "flood";
-  if (normalized.includes("volcano")) return "volcano";
-  if (normalized.includes("ice")) return "ice";
-  if (normalized.includes("drought")) return "drought";
-  if (normalized.includes("dust")) return "dust";
-  return "default";
-};
-
-type OverlayBadge = {
-  label: string;
-  icon?: React.ComponentType<{ className?: string }>;
-  className?: string;
-};
-
-type OverlayDetailItem = {
-  label: string;
-  value: string;
-  icon?: React.ComponentType<{ className?: string }>;
-  cardClassName?: string;
-  labelClassName?: string;
-};
-
-type OverlayAction = {
-  label: string;
-  url: string;
-  ariaLabel: string;
-};
-
-const isDefined = <T,>(value: T | null | undefined): value is T => value !== null && value !== undefined;
-
-const OverlayDetailGrid = ({
-  details,
-  className,
-}: {
-  details: OverlayDetailItem[];
-  className?: string;
-}) => {
-  if (details.length === 0) return null;
-
-  return (
-    <div className={cn("grid grid-cols-1 gap-2 sm:grid-cols-2", className)}>
-      {details.map((detail) => (
-        <div
-          key={detail.label}
-          className={cn("rounded-lg bg-muted/40 px-3 py-2 dark:bg-muted/20", detail.cardClassName)}
-        >
-          <div
-            className={cn(
-              "flex items-center gap-2 text-[10px] uppercase tracking-wide text-muted-foreground",
-              detail.labelClassName,
-            )}
-          >
-            {detail.icon ? <detail.icon className="size-3" /> : null}
-            <span>{detail.label}</span>
-          </div>
-          <p className="text-sm font-medium text-foreground">{detail.value}</p>
-        </div>
-      ))}
-    </div>
-  );
-};
-
-const getEventTone = (category: string) => {
-  const toneKey = getEventToneKey(category);
-  if (toneKey === "wildfire") {
-    return {
-      lead: "bg-orange-500/20 text-orange-700 dark:bg-orange-500/30 dark:text-orange-200",
-      badge: "bg-orange-500/15 text-orange-700 dark:bg-orange-500/25 dark:text-orange-200",
-      detailCard: "bg-orange-500/8 border border-orange-500/20 dark:bg-orange-500/12 dark:border-orange-400/20",
-      detailLabel: "text-orange-700/80 dark:text-orange-200/80",
-    };
-  }
-  if (toneKey === "storm") {
-    return {
-      lead: "bg-amber-500/20 text-amber-700 dark:bg-amber-500/30 dark:text-amber-200",
-      badge: "bg-amber-500/15 text-amber-700 dark:bg-amber-500/25 dark:text-amber-200",
-      detailCard: "bg-amber-500/8 border border-amber-500/20 dark:bg-amber-500/12 dark:border-amber-400/20",
-      detailLabel: "text-amber-700/80 dark:text-amber-200/80",
-    };
-  }
-  if (toneKey === "flood") {
-    return {
-      lead: "bg-sky-500/20 text-sky-700 dark:bg-sky-500/30 dark:text-sky-200",
-      badge: "bg-sky-500/15 text-sky-700 dark:bg-sky-500/25 dark:text-sky-200",
-      detailCard: "bg-sky-500/8 border border-sky-500/20 dark:bg-sky-500/12 dark:border-sky-400/20",
-      detailLabel: "text-sky-700/80 dark:text-sky-200/80",
-    };
-  }
-  if (toneKey === "volcano") {
-    return {
-      lead: "bg-pink-500/20 text-pink-700 dark:bg-pink-500/30 dark:text-pink-200",
-      badge: "bg-pink-500/15 text-pink-700 dark:bg-pink-500/25 dark:text-pink-200",
-      detailCard: "bg-pink-500/8 border border-pink-500/20 dark:bg-pink-500/12 dark:border-pink-400/20",
-      detailLabel: "text-pink-700/80 dark:text-pink-200/80",
-    };
-  }
-  if (toneKey === "ice") {
-    return {
-      lead: "bg-cyan-500/20 text-cyan-700 dark:bg-cyan-500/30 dark:text-cyan-200",
-      badge: "bg-cyan-500/15 text-cyan-700 dark:bg-cyan-500/25 dark:text-cyan-200",
-      detailCard: "bg-cyan-500/8 border border-cyan-500/20 dark:bg-cyan-500/12 dark:border-cyan-400/20",
-      detailLabel: "text-cyan-700/80 dark:text-cyan-200/80",
-    };
-  }
-  if (toneKey === "drought") {
-    return {
-      lead: "bg-yellow-700/20 text-yellow-700 dark:bg-yellow-700/30 dark:text-yellow-200",
-      badge: "bg-yellow-700/15 text-yellow-700 dark:bg-yellow-700/25 dark:text-yellow-200",
-      detailCard: "bg-yellow-700/8 border border-yellow-700/20 dark:bg-yellow-700/12 dark:border-yellow-500/20",
-      detailLabel: "text-yellow-700/80 dark:text-yellow-200/80",
-    };
-  }
-  if (toneKey === "dust") {
-    return {
-      lead: "bg-yellow-500/20 text-yellow-700 dark:bg-yellow-500/30 dark:text-yellow-200",
-      badge: "bg-yellow-500/15 text-yellow-700 dark:bg-yellow-500/25 dark:text-yellow-200",
-      detailCard: "bg-yellow-500/8 border border-yellow-500/20 dark:bg-yellow-500/12 dark:border-yellow-400/20",
-      detailLabel: "text-yellow-700/80 dark:text-yellow-200/80",
-    };
-  }
-  return {
-    lead: "bg-amber-500/20 text-amber-700 dark:bg-amber-500/30 dark:text-amber-200",
-    badge: "bg-amber-500/15 text-amber-700 dark:bg-amber-500/25 dark:text-amber-200",
-    detailCard: "bg-amber-500/8 border border-amber-500/20 dark:bg-amber-500/12 dark:border-amber-400/20",
-    detailLabel: "text-amber-700/80 dark:text-amber-200/80",
-  };
-};
-
-const UnifiedSignalPopover = ({
-  leading,
-  title,
-  titlePrefixIcon: TitlePrefixIcon,
-  onClose,
-  badges,
-  details,
-  detailsClassName,
-  footerAction,
-  children,
-}: {
-  leading: React.ReactNode;
-  title: string;
-  titlePrefixIcon?: React.ComponentType<{ className?: string }>;
-  onClose: () => void;
-  badges?: OverlayBadge[];
-  details?: OverlayDetailItem[];
-  detailsClassName?: string;
-  footerAction?: OverlayAction | null;
-  children?: React.ReactNode;
-}) => {
-  return (
-    <div className="space-y-4">
-      <div className="flex items-start gap-3">
-        {leading}
-        <div className="min-w-0 flex-1 space-y-2">
-          <div className="flex items-start justify-between gap-3">
-            <div className="flex min-w-0 items-center gap-1 text-sm text-muted-foreground">
-              {TitlePrefixIcon ? <TitlePrefixIcon className="size-3.5 shrink-0" /> : null}
-              <p className="line-clamp-2 text-foreground font-semibold leading-snug">{title}</p>
-            </div>
-            <button
-              type="button"
-              onClick={onClose}
-              className="shrink-0 rounded-md p-1 text-muted-foreground hover:bg-muted/70 hover:text-foreground transition-colors"
-              aria-label="Close popover"
-            >
-              <X className="size-4" />
-            </button>
-          </div>
-          {badges && badges.length > 0 ? (
-            <div className="flex flex-wrap gap-2">
-              {badges.map((badge) => (
-                <Badge key={badge.label} className={badge.className}>
-                  {badge.icon ? <badge.icon className="size-3" /> : null}
-                  {badge.label}
-                </Badge>
-              ))}
-            </div>
-          ) : null}
-        </div>
-      </div>
-
-      {details ? <OverlayDetailGrid details={details} className={detailsClassName} /> : null}
-      {children}
-
-      {footerAction ? (
-        <Button
-          asChild
-          className="w-full gap-2 rounded-lg bg-primary/10 hover:bg-primary/20 text-primary px-4 py-2.5 text-sm font-medium transition-colors"
-        >
-          <a
-            href={footerAction.url}
-            target="_blank"
-            rel="noopener noreferrer"
-            aria-label={footerAction.ariaLabel}
-          >
-            {footerAction.label}
-            <ExternalLink className="size-4" />
-          </a>
-        </Button>
-      ) : null}
-    </div>
-  );
-};
-
-export function MapStateSync({
-  setViewState,
-  onUserInteract,
-  viewStateKey,
-  viewStateSourceKey,
-}: {
-  setViewState: (s: MapViewState) => void;
-  onUserInteract?: () => void;
-  viewStateKey: string;
-  viewStateSourceKey: string;
-}) {
-  const { map } = useMap();
-  const hasUserInteractedRef = React.useRef(false);
-
-  React.useEffect(() => {
-    if (!map) return;
-
-    const markUserInteraction = () => {
-      hasUserInteractedRef.current = true;
-      onUserInteract?.();
-    };
-
-    const handleMoveEnd = () => {
-      if (!hasUserInteractedRef.current) return;
-      const newState = {
-        center: [map.getCenter().lng, map.getCenter().lat] as [number, number],
-        zoom: map.getZoom(),
-      };
-      setViewState(newState);
-      try {
-        localStorage.setItem(viewStateKey, JSON.stringify(newState));
-        localStorage.setItem(viewStateSourceKey, "user");
-      } catch {
-        // ignore storage errors
-      }
-    };
-
-    map.on("dragstart", markUserInteraction);
-    map.on("zoomstart", markUserInteraction);
-    map.on("rotatestart", markUserInteraction);
-    map.on("pitchstart", markUserInteraction);
-    map.on("moveend", handleMoveEnd);
-    return () => {
-      map.off("dragstart", markUserInteraction);
-      map.off("zoomstart", markUserInteraction);
-      map.off("rotatestart", markUserInteraction);
-      map.off("pitchstart", markUserInteraction);
-      map.off("moveend", handleMoveEnd);
-    };
-  }, [map, setViewState, onUserInteract, viewStateKey, viewStateSourceKey]);
-
-  return null;
-}
-
-export function MapViewController({
-  viewState,
-  shouldAutoCenter,
-}: {
-  viewState: MapViewState;
-  shouldAutoCenter: boolean;
-}) {
-  const { map, isLoaded } = useMap();
-  const lastAppliedRef = React.useRef<MapViewState | null>(null);
-
-  React.useEffect(() => {
-    if (!map || !isLoaded || !shouldAutoCenter) return;
-
-    const lastApplied = lastAppliedRef.current;
-    if (
-      lastApplied &&
-      lastApplied.center[0] === viewState.center[0] &&
-      lastApplied.center[1] === viewState.center[1] &&
-      lastApplied.zoom === viewState.zoom
-    ) {
-      return;
-    }
-
-    map.easeTo({
-      center: viewState.center,
-      zoom: viewState.zoom,
-      duration: 900,
-      essential: true,
-    });
-    lastAppliedRef.current = viewState;
-  }, [map, isLoaded, shouldAutoCenter, viewState]);
-
-  return null;
-}
-
-export function EarthquakeFlyTo({
-  earthquake,
-}: {
-  earthquake: ProcessedEarthquake | null;
-}) {
-  const { map } = useMap();
-  const prevEarthquakeId = React.useRef<string | null>(null);
-
-  React.useEffect(() => {
-    if (!map || !earthquake) return;
-
-    if (prevEarthquakeId.current === earthquake.id) return;
-    prevEarthquakeId.current = earthquake.id;
-
-    const easeOut = (t: number) => 1 - Math.pow(1 - t, 3);
-    map.stop();
-    map.easeTo({
-      center: [earthquake.coordinates[0], earthquake.coordinates[1]],
-      zoom: Math.max(map.getZoom(), 8),
-      duration: 900,
-      easing: easeOut,
-      essential: true,
-    });
-  }, [map, earthquake]);
-
-  return null;
-}
-
-export function EonetFlyTo({
-  event,
-}: {
-  event: ProcessedEonetEvent | null;
-}) {
-  const { map } = useMap();
-
-  React.useEffect(() => {
-    if (!map || !event) return;
-
-    map.stop();
-    map.easeTo({
-      center: event.coordinates,
-      zoom: Math.max(map.getZoom(), 5.8),
-      duration: 900,
-      essential: true,
-    });
-  }, [map, event]);
-
-  return null;
-}
-
-export function TsunamiFlyTo({
-  alert,
-}: {
-  alert: ProcessedTsunamiAlert | null;
-}) {
-  const { map } = useMap();
-
-  React.useEffect(() => {
-    if (!map || !alert) return;
-
-    map.stop();
-    map.easeTo({
-      center: alert.coordinates,
-      zoom: Math.max(map.getZoom(), 5.8),
-      duration: 900,
-      essential: true,
-    });
-  }, [map, alert]);
-
-  return null;
-}
-
-export function AirQualityFlyTo({
-  site,
-}: {
-  site: ProcessedAirQualitySite | null;
-}) {
-  const { map } = useMap();
-
-  React.useEffect(() => {
-    if (!map || !site) return;
-
-    map.stop();
-    map.easeTo({
-      center: site.coordinates,
-      zoom: Math.max(map.getZoom(), 6),
-      duration: 900,
-      essential: true,
-    });
-  }, [map, site]);
-
-  return null;
-}
-
-export function SignalOverlay({
-  activeType,
-  earthquake,
-  event,
-  tsunamiAlert,
-  airQualitySite,
-  onCloseEarthquake,
-  onCloseEvent,
-}: {
-  activeType: "earthquake" | "global" | null;
-  earthquake: ProcessedEarthquake | null;
-  event: ProcessedEonetEvent | null;
-  tsunamiAlert: ProcessedTsunamiAlert | null;
-  airQualitySite: ProcessedAirQualitySite | null;
-  onCloseEarthquake: () => void;
-  onCloseEvent: () => void;
-}) {
-  const { map } = useMap();
-  const overlayContainer = React.useMemo(() => document.createElement("div"), []);
-
-  React.useEffect(() => {
-    if (!map) return;
-    const container = map.getContainer();
-    overlayContainer.className =
-      "absolute left-3 right-3 top-3 z-30 pointer-events-auto sm:left-auto sm:right-4 sm:top-4";
-    container.appendChild(overlayContainer);
-    return () => {
-      overlayContainer.remove();
-    };
-  }, [map, overlayContainer]);
-
-  const activeEarthquake = activeType === "earthquake" ? earthquake : null;
-  const activeEvent = activeType === "global" ? event : null;
-  const activeTsunami = activeType === "global" ? tsunamiAlert : null;
-  const activeAirQuality = activeType === "global" ? airQualitySite : null;
-
-  if (!map || (!activeEarthquake && !activeEvent && !activeTsunami && !activeAirQuality)) return null;
-
-  const formatRelativeTime = (date: Date): string => {
-    const now = new Date();
-    const diffMs = now.getTime() - date.getTime();
-    const diffMins = Math.floor(diffMs / 60000);
-    const diffHours = Math.floor(diffMins / 60);
-    const diffDays = Math.floor(diffHours / 24);
-    if (diffMins < 1) return "Just now";
-    if (diffMins < 60) return `${diffMins}m ago`;
-    if (diffHours < 24) return `${diffHours}h ago`;
-    return `${diffDays}d ago`;
-  };
-
-  const detailItems: OverlayDetailItem[] = activeEarthquake
-    ? (
-        [
-          {
-            label: "Magnitude",
-            value: `${activeEarthquake.magnitude.toFixed(1)} (${getMagnitudeLabel(activeEarthquake.magnitude)})`,
-            icon: Gauge,
-          },
-          activeEarthquake.magType
-            ? {
-                label: "Magnitude scale",
-                value: activeEarthquake.magType.toUpperCase(),
-                icon: Gauge,
-              }
-            : null,
-          { label: "Depth below ground", value: `${activeEarthquake.depth.toFixed(1)} km`, icon: MapPin },
-          { label: "Impact score", value: `${activeEarthquake.sig}`, icon: Signal },
-          activeEarthquake.felt !== null
-            ? { label: "Felt", value: `${activeEarthquake.felt}`, icon: Users }
-            : null,
-          activeEarthquake.mmi !== null
-            ? { label: "Intensity (MMI)", value: activeEarthquake.mmi.toFixed(1), icon: Gauge }
-            : null,
-          activeEarthquake.cdi !== null
-            ? { label: "Community intensity", value: activeEarthquake.cdi.toFixed(1), icon: Signal }
-            : null,
-          activeEarthquake.gap !== null
-            ? { label: "Azimuthal gap", value: `${activeEarthquake.gap.toFixed(0)}°`, icon: CircleDot }
-            : null,
-          activeEarthquake.dmin !== null
-            ? { label: "Nearest station", value: activeEarthquake.dmin.toFixed(2), icon: MapPin }
-            : null,
-          activeEarthquake.rms !== null
-            ? { label: "Wave residual", value: activeEarthquake.rms.toFixed(2), icon: Radio }
-            : null,
-          activeEarthquake.nst !== null
-            ? { label: "Stations reporting", value: `${activeEarthquake.nst}`, icon: Radio }
-            : null,
-        ].filter(isDefined)
-      )
-    : [];
-
-  const earthquakeBadges: OverlayBadge[] = activeEarthquake
-    ? [
-        {
-          label: `Updated ${formatRelativeTime(activeEarthquake.updated)}`,
-          icon: Clock,
-          className: "gap-1 bg-sky-500/20 text-sky-700 dark:bg-sky-500/30 dark:text-sky-200",
-        },
-        activeEarthquake.alert
-          ? {
-              label: `Alert ${activeEarthquake.alert}`,
-              icon: AlertTriangle,
-              className: cn(
-                "gap-1",
-                activeEarthquake.alert === "yellow" &&
-                  "bg-yellow-500/20 text-yellow-700 dark:bg-yellow-500/30 dark:text-yellow-200",
-                activeEarthquake.alert === "orange" &&
-                  "bg-orange-500/20 text-orange-700 dark:bg-orange-500/30 dark:text-orange-200",
-                activeEarthquake.alert === "red" &&
-                  "bg-red-500/20 text-red-700 dark:bg-red-500/30 dark:text-red-200",
-                activeEarthquake.alert === "green" &&
-                  "bg-emerald-500/20 text-emerald-700 dark:bg-emerald-500/30 dark:text-emerald-200",
-              ),
-            }
-          : null,
-        activeEarthquake.tsunami
-          ? {
-              label: "Tsunami risk",
-              icon: Waves,
-              className: "gap-1 bg-blue-500/20 text-blue-700 dark:bg-blue-500/30 dark:text-blue-200",
-            }
-          : null,
-      ].filter(isDefined)
-    : [];
-
-  const eventDetails: OverlayDetailItem[] = activeEvent
-    ? [
-        {
-          label: "Coordinates",
-          value: `${activeEvent.coordinates[1].toFixed(4)}, ${activeEvent.coordinates[0].toFixed(4)}`,
-          icon: MapPinned,
-        },
-        {
-          label: "Reported",
-          value: activeEvent.date.toLocaleString("en-US", {
-            month: "short",
-            day: "numeric",
-            hour: "numeric",
-            minute: "2-digit",
-          }),
-          icon: CalendarClock,
-        },
-      ]
-    : [];
-
-  const eventBadges: OverlayBadge[] = activeEvent
-    ? [
-        {
-          label: getEventCategoryLabel(activeEvent.category),
-          icon: activeEvent ? getEventIcon(activeEvent.category) : Globe2,
-          className: cn("gap-1", getEventTone(activeEvent.category).badge),
-        },
-      ]
-    : [];
-
-  const tsunamiDetails: OverlayDetailItem[] = activeTsunami
-    ? [
-        {
-          label: "Coordinates",
-          value: `${activeTsunami.coordinates[1].toFixed(4)}, ${activeTsunami.coordinates[0].toFixed(4)}`,
-          icon: MapPinned,
-        },
-      ]
-    : [];
-
-  const tsunamiBadges: OverlayBadge[] = activeTsunami
-    ? [
-        {
-          label: activeTsunami.severity,
-          className: "bg-sky-500/15 text-sky-700 dark:bg-sky-500/25 dark:text-sky-200",
-        },
-          activeTsunami.sent
-            ? {
-                label: activeTsunami.sent.toLocaleString("en-US", {
-                  month: "short",
-                  day: "numeric",
-                  hour: "numeric",
-                  minute: "2-digit",
-                }),
-                icon: CalendarClock,
-                className: "bg-slate-500/15 text-slate-700 dark:bg-slate-500/25 dark:text-slate-200",
-              }
-            : null,
-      ].filter(isDefined)
-    : [];
-
-  const airQualityBadges: OverlayBadge[] = activeAirQuality
-    ? [
-        {
-          label: formatAirQualityParameterLabel(activeAirQuality.parameter),
-          className: "bg-emerald-500/15 text-emerald-700 dark:bg-emerald-500/25 dark:text-emerald-200",
-        },
-        {
-          label: `Concentration ${formatAirQualityConcentration(activeAirQuality.value, activeAirQuality.unit)}`,
-          className: "bg-slate-500/15 text-slate-700 dark:bg-slate-500/25 dark:text-slate-200",
-        },
-      ]
-    : [];
-
-  const airQualityDetails: OverlayDetailItem[] = activeAirQuality
-    ? (
-        [
-          {
-            label: "Coordinates",
-            value: `${activeAirQuality.coordinates[1].toFixed(4)}, ${activeAirQuality.coordinates[0].toFixed(4)}`,
-            icon: MapPinned,
-          },
-          activeAirQuality.measuredAt
-            ? {
-                label: "Measured",
-                value: activeAirQuality.measuredAt.toLocaleString("en-US", {
-                  month: "short",
-                  day: "numeric",
-                  hour: "numeric",
-                  minute: "2-digit",
-                }),
-                icon: CalendarClock,
-              }
-            : null,
-          activeAirQuality.averagingPeriod
-            ? {
-                label: "Averaging period",
-                value: activeAirQuality.averagingPeriod,
-                icon: ScanLine,
-              }
-            : null,
-          activeAirQuality.coveragePercent !== null && activeAirQuality.coveragePercent !== undefined
-            ? {
-                label: "Data coverage",
-                value: `${activeAirQuality.coveragePercent.toFixed(0)}%`,
-                icon: Signal,
-              }
-            : null,
-          {
-            label: "Location ID",
-            value: `${activeAirQuality.locationId}`,
-            icon: Building2,
-          },
-          {
-            label: "Sensor ID",
-            value: `${activeAirQuality.sensorId}`,
-            icon: Radio,
-          },
-          activeAirQuality.city || activeAirQuality.country
-            ? {
-                label: "Region",
-                value: [activeAirQuality.city, activeAirQuality.country].filter(Boolean).join(", "),
-                icon: MapPin,
-              }
-            : null,
-          activeAirQuality.averageValue !== null && activeAirQuality.averageValue !== undefined
-            ? {
-                label: "Average",
-                value: formatAirQualityConcentration(activeAirQuality.averageValue, activeAirQuality.unit),
-                icon: Ruler,
-              }
-            : null,
-          activeAirQuality.minValue !== null && activeAirQuality.minValue !== undefined
-            ? {
-                label: "Minimum",
-                value: formatAirQualityConcentration(activeAirQuality.minValue, activeAirQuality.unit),
-                icon: FlaskConical,
-              }
-            : null,
-          activeAirQuality.maxValue !== null && activeAirQuality.maxValue !== undefined
-            ? {
-                label: "Maximum",
-                value: formatAirQualityConcentration(activeAirQuality.maxValue, activeAirQuality.unit),
-                icon: FlaskConical,
-              }
-            : null,
-        ].filter(isDefined)
-      )
-    : [];
-
-  const ActiveEventIcon = activeEvent ? getEventIcon(activeEvent.category) : Globe2;
-  const activeEventTone = activeEvent ? getEventTone(activeEvent.category) : getEventTone("default");
-
-  return createPortal(
-    <div className="w-full max-w-sm sm:max-w-md max-h-[75vh] overflow-y-auto rounded-lg border border-border/60 bg-background/95 p-3">
-      {activeEarthquake ? (
-        <UnifiedSignalPopover
-          leading={(
-            <div
-              className="flex size-12 shrink-0 items-center justify-center rounded-lg text-lg font-bold text-white"
-              style={{
-                backgroundColor: getMagnitudeColor(activeEarthquake.magnitude),
-                boxShadow: `0 4px 14px ${getMagnitudeColor(activeEarthquake.magnitude)}40`,
-              }}
-            >
-              {activeEarthquake.magnitude.toFixed(1)}
-            </div>
-          )}
-          title={activeEarthquake.place}
-          titlePrefixIcon={MapPin}
-          onClose={onCloseEarthquake}
-          badges={earthquakeBadges}
-          details={detailItems}
-          detailsClassName="grid-cols-2 sm:grid-cols-2"
-          footerAction={{
-            label: "View on USGS",
-            url: activeEarthquake.url,
-            ariaLabel: "View USGS event details",
-          }}
-        >
-          <div className="flex flex-wrap gap-2">
-            <Badge className="gap-1 bg-slate-500/15 text-slate-700 dark:bg-slate-500/25 dark:text-slate-200">
-              <MapPin className="size-3" />
-              Latitude: {activeEarthquake.coordinates[1].toFixed(4)}
-            </Badge>
-            <Badge className="gap-1 bg-slate-500/15 text-slate-700 dark:bg-slate-500/25 dark:text-slate-200">
-              <MapPin className="size-3" />
-              Longitude: {activeEarthquake.coordinates[0].toFixed(4)}
-            </Badge>
-          </div>
-        </UnifiedSignalPopover>
-      ) : null}
-
-      {activeEvent ? (
-        <UnifiedSignalPopover
-          leading={(
-            <div className={cn("flex size-12 shrink-0 items-center justify-center rounded-lg", activeEventTone.lead)}>
-              <ActiveEventIcon className="size-6" />
-            </div>
-          )}
-          title={activeEvent.title}
-          onClose={onCloseEvent}
-          badges={eventBadges}
-          details={eventDetails}
-        />
-      ) : null}
-
-      {activeTsunami ? (
-        <UnifiedSignalPopover
-          leading={(
-            <div className="flex size-12 shrink-0 items-center justify-center rounded-lg bg-sky-500/20 text-sky-700 dark:bg-sky-500/30 dark:text-sky-200">
-              <Waves className="size-6" />
-            </div>
-          )}
-          title={activeTsunami.headline}
-          onClose={onCloseEvent}
-          badges={tsunamiBadges}
-          details={tsunamiDetails}
-        />
-      ) : null}
-
-      {activeAirQuality ? (
-        <UnifiedSignalPopover
-          leading={(
-            <div className="flex size-12 shrink-0 items-center justify-center rounded-lg bg-emerald-500/20 text-emerald-700 dark:bg-emerald-500/30 dark:text-emerald-200">
-              <Wind className="size-6" />
-            </div>
-          )}
-          title={activeAirQuality.location}
-          onClose={onCloseEvent}
-          badges={airQualityBadges}
-          details={airQualityDetails}
-        />
-      ) : null}
-    </div>,
-    overlayContainer
-  );
-}
-
-type ControlButtonProps = Omit<
-  React.ComponentPropsWithoutRef<"button">,
-  "children" | "aria-label"
-> & {
-  label: string;
-  children: React.ReactNode;
-  active?: boolean;
-};
-
-const ControlButton = React.forwardRef<HTMLButtonElement, ControlButtonProps>(
-  function ControlButton(
-    { label, children, active = false, className, type, disabled, ...props },
-    ref,
-  ) {
-    return (
-      <button
-        ref={ref}
-        type={type ?? "button"}
-        aria-label={label}
-        className={cn(
-          "flex items-center justify-center size-8 transition-colors",
-          active
-            ? "bg-primary text-primary-foreground hover:bg-primary/90"
-            : "hover:bg-accent hover:text-accent-foreground text-foreground",
-          disabled && "opacity-50 pointer-events-none cursor-not-allowed",
-          className,
-        )}
-        disabled={disabled}
-        {...props}
-      >
-        {children}
-      </button>
-    );
-  },
-);
-
-function ControlGroup({ children }: { children: React.ReactNode }) {
-  return (
-    <div className="flex flex-col rounded-md border border-border bg-background shadow-sm overflow-hidden [&>button:not(:last-child)]:border-b [&>button:not(:last-child)]:border-border">
-      {children}
-    </div>
-  );
-}
-
-function MobileDrawerHeader({
-  icon: Icon,
-  title,
-  description,
-  iconToneClassName,
-  onClose,
-}: {
-  icon: React.ComponentType<{ className?: string }>;
-  title: string;
-  description: string;
-  iconToneClassName: string;
-  onClose: () => void;
-}) {
-  return (
-    <div className="border-b border-border/60 px-4 pb-3 pt-2">
-      <div className="flex items-start justify-between gap-3">
-        <div className="flex min-w-0 items-center gap-2.5">
-          <div
-            className={cn(
-              "flex size-8 shrink-0 items-center justify-center rounded-md border border-white/20",
-              iconToneClassName,
-            )}
-          >
-            <Icon className="size-4 text-white" />
-          </div>
-          <div className="min-w-0">
-            <h3 className="truncate text-sm font-semibold">{title}</h3>
-            <p className="truncate text-xs text-muted-foreground">{description}</p>
-          </div>
-        </div>
-        <Button
-          type="button"
-          size="icon-sm"
-          variant="ghost"
-          className="shrink-0 rounded-md border border-border/60 text-muted-foreground hover:bg-muted/60 hover:text-foreground"
-          onClick={onClose}
-          aria-label={`Close ${title} drawer`}
-        >
-          <X className="size-4" />
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-export function MapOverlayUI({
-  setUserLocation,
-  onLocateAnimation,
-  resolvedLocation,
-  earthquakes,
-  onEarthquakeSelect,
-  onEonetSelect,
-  onAirQualitySelect,
-  onTsunamiSelect,
-  appSettings,
-  onAppSettingsChange,
-  onResetDefaults,
-  eonetState,
-  airQualityState,
-  tsunamiState,
-  layerVisibility,
-  onLayerVisibilityChange,
-}: {
-  setUserLocation: (l: [number, number]) => void;
-  onLocateAnimation: () => void;
-  resolvedLocation: [number, number] | null;
-  earthquakes: ProcessedEarthquake[];
-  onEarthquakeSelect: (eq: ProcessedEarthquake) => void;
-  onEonetSelect: (event: ProcessedEonetEvent) => void;
-  onAirQualitySelect: (site: ProcessedAirQualitySite) => void;
-  onTsunamiSelect: (alert: ProcessedTsunamiAlert) => void;
-  appSettings: AppSettings;
-  onAppSettingsChange: React.Dispatch<React.SetStateAction<AppSettings>>;
-  onResetDefaults: () => void;
-  eonetState: {
-    events: ProcessedEonetEvent[];
-    isLoading: boolean;
-    error: Error | null;
-    lastUpdated: Date | null;
-    refetch: () => Promise<void>;
-  };
-  airQualityState: {
-    sites: ProcessedAirQualitySite[];
-    isLoading: boolean;
-    error: Error | null;
-    lastUpdated: Date | null;
-    refetch: () => Promise<void>;
-  };
-  tsunamiState: {
-    alerts: ProcessedTsunamiAlert[];
-    isLoading: boolean;
-    error: Error | null;
-    lastUpdated: Date | null;
-    refetch: () => Promise<void>;
-  };
-  layerVisibility: LayerVisibility;
-  onLayerVisibilityChange: React.Dispatch<React.SetStateAction<LayerVisibility>>;
-}) {
-  const [activeMobileDrawer, setActiveMobileDrawer] = React.useState<"menu" | "quakes" | "weather" | "global" | null>(null);
-  const [isMap3DMode, setIsMap3DMode] = React.useState(false);
-  const [isMapGlobeMode, setIsMapGlobeMode] = React.useState(false);
-  const { is3D, toggle3D, isGlobe, toggleGlobe } = useMap3DMode(setIsMap3DMode, setIsMapGlobeMode);
-  const isMobile = useIsMobile();
-  const [quakePage, setQuakePage] = React.useState(1);
-  const quakePageSize = 8;
-  const totalQuakePages = Math.max(1, Math.ceil(earthquakes.length / quakePageSize));
-  const paginatedEarthquakes = React.useMemo(() => {
-    const startIndex = (quakePage - 1) * quakePageSize;
-    return earthquakes.slice(startIndex, startIndex + quakePageSize);
-  }, [earthquakes, quakePage]);
-  const quakePaginationRange = React.useMemo(
-    () => getPaginationRange(totalQuakePages, quakePage),
-    [quakePage, totalQuakePages]
-  );
-
-  React.useEffect(() => {
-    setQuakePage(1);
-  }, [earthquakes]);
-
-  React.useEffect(() => {
-    if (isMobile) return;
-    setActiveMobileDrawer(null);
-  }, [isMobile]);
-
-  const activateDrawer = React.useCallback((drawer: "menu" | "quakes" | "weather" | "global") => {
-    setActiveMobileDrawer(drawer);
-  }, []);
-
-  const closeDrawer = React.useCallback(() => {
-    setActiveMobileDrawer(null);
-  }, []);
-
-  const handleQuakeClick = (eq: ProcessedEarthquake) => {
-    closeDrawer();
-    onEarthquakeSelect(eq);
-  };
-
-  const handlePreviousQuakePage = () => {
-    setQuakePage((page) => Math.max(1, page - 1));
-  };
-
-  const handleNextQuakePage = () => {
-    setQuakePage((page) => Math.min(totalQuakePages, page + 1));
-  };
-
-  return (
-    <div className="absolute inset-0 pointer-events-none">
-      <div className="pointer-events-auto absolute top-[calc(env(safe-area-inset-top)+1rem)] sm:top-4 left-3 sm:left-4 z-10">
-        <MapCameraDock is3DModeEnabled={isMap3DMode || isMapGlobeMode} />
-      </div>
-
-      <div className="pointer-events-auto absolute bottom-[calc(env(safe-area-inset-bottom)+5.5rem)] sm:bottom-4 left-3 sm:left-4 z-20">
-        <MapLayersControl 
-          appSettings={appSettings} 
-          onAppSettingsChange={onAppSettingsChange} 
-          is3D={is3D} 
-          toggle3D={toggle3D} 
-          isGlobe={isGlobe}
-          toggleGlobe={toggleGlobe}
-        />
-      </div>
-
-      <div className="pointer-events-auto absolute bottom-[calc(env(safe-area-inset-bottom)+5.5rem)] sm:bottom-4 right-3 sm:right-4 z-10">
-        <CustomMapControls
-          setUserLocation={setUserLocation}
-          onLocateAnimation={onLocateAnimation}
-          layerVisibility={layerVisibility}
-          onLayerVisibilityChange={onLayerVisibilityChange}
-        />
-      </div>
-
-      {appSettings.showDesktopForecastDock ? (
-        <div className="pointer-events-none absolute inset-x-0 bottom-4 hidden md:flex justify-center z-10">
-          <div className="pointer-events-auto">
-            <HourlyForecastDock
-              latitude={resolvedLocation?.[1] ?? null}
-              longitude={resolvedLocation?.[0] ?? null}
-              temperatureUnit={appSettings.temperatureUnit}
-              timeFormat={appSettings.timeFormat}
-            />
-          </div>
-        </div>
-      ) : null}
-
-      <Drawer
-        open={activeMobileDrawer !== null}
-        onOpenChange={(open) => {
-          if (!open) {
-            closeDrawer();
-          }
-        }}
-      >
-        <DrawerContent className="h-[88dvh] max-h-[88dvh] min-h-0 rounded-t-2xl border-t border-border/70 p-0">
-          <div className="flex h-full min-h-0 flex-col">
-            {activeMobileDrawer === "menu" ? (
-              <>
-                <MobileDrawerHeader
-                  icon={SlidersHorizontal}
-                  iconToneClassName="bg-slate-500"
-                  title="Menu"
-                  description="Settings and layer controls"
-                  onClose={closeDrawer}
-                />
-                <div
-                  data-vaul-no-drag
-                  className="min-h-0 flex-1 overflow-y-auto overscroll-contain px-3 pb-3"
-                >
-                  <div className="rounded-md border border-border/60 bg-muted/15 p-3">
-                    <HazrSettingsPanel
-                      settings={appSettings}
-                      onSettingsChange={onAppSettingsChange}
-                      layerVisibility={layerVisibility}
-                      onLayerVisibilityChange={onLayerVisibilityChange}
-                      onResetDefaults={onResetDefaults}
-                      showInlineReset
-                    />
-                  </div>
-                  <div className="mt-3 rounded-md border border-border/60 bg-muted/15 p-3">
-                    <div className="flex items-center gap-2 mb-3">
-                      <span className="inline-flex items-center justify-center rounded-md bg-blue-500/15 text-white">
-                        <Map className="size-5" />
-                      </span>
-                      <h3 className="text-sm font-semibold">About Hazr</h3>
-                    </div>
-                    <HazrAboutContent />
-                  </div>
-                </div>
-              </>
-            ) : null}
-
-            {activeMobileDrawer === "quakes" ? (
-              <>
-                <MobileDrawerHeader
-                  icon={Activity}
-                  iconToneClassName="bg-red-500"
-                  title="Live Earthquakes"
-                  description={`${earthquakes.length} in the last 24h`}
-                  onClose={closeDrawer}
-                />
-                <div className="flex min-h-0 flex-1 flex-col px-3 pb-3">
-                  {!layerVisibility.earthquakes ? (
-                    <div className="flex min-h-0 flex-1 items-center justify-center rounded-md border border-border/60 bg-muted/15 px-3 py-6 text-center text-sm text-muted-foreground">
-                      USGS earthquakes are hidden in settings.
-                    </div>
-                  ) : earthquakes.length === 0 ? (
-                    <div className="flex min-h-0 flex-1 items-center justify-center rounded-md border border-border/60 bg-muted/15 px-3 py-6 text-center text-sm text-muted-foreground">
-                      No recent earthquakes
-                    </div>
-                  ) : (
-                    <>
-                      <div
-                        data-vaul-no-drag
-                        className="min-h-0 flex-1 overflow-y-auto overscroll-contain rounded-md border border-border/60 bg-muted/15 p-1.5"
-                      >
-                        <div className="flex flex-col gap-1">
-                          {paginatedEarthquakes.map((eq) => (
-                            <EarthquakeItem
-                              key={eq.id}
-                              earthquake={eq}
-                              onClick={() => handleQuakeClick(eq)}
-                            />
-                          ))}
-                        </div>
-                      </div>
-                      {totalQuakePages > 1 ? (
-                        <Pagination className="mt-2">
-                          <PaginationContent className="hidden sm:flex">
-                            <PaginationItem>
-                              <PaginationPrevious
-                                onClick={handlePreviousQuakePage}
-                                disabled={quakePage === 1}
-                                aria-disabled={quakePage === 1}
-                              />
-                            </PaginationItem>
-                            {quakePaginationRange.map((entry, index) => (
-                              <PaginationItem key={`mobile-quake-page-${entry}-${index}`}>
-                                {entry === "ellipsis" ? (
-                                  <PaginationEllipsis />
-                                ) : (
-                                  <PaginationButton
-                                    onClick={() => setQuakePage(entry)}
-                                    isActive={entry === quakePage}
-                                    aria-label={`Go to quake page ${entry}`}
-                                  >
-                                    {entry}
-                                  </PaginationButton>
-                                )}
-                              </PaginationItem>
-                            ))}
-                            <PaginationItem>
-                              <PaginationNext
-                                onClick={handleNextQuakePage}
-                                disabled={quakePage === totalQuakePages}
-                                aria-disabled={quakePage === totalQuakePages}
-                              />
-                            </PaginationItem>
-                          </PaginationContent>
-                          <PaginationContent className="flex w-full items-center justify-between gap-2 px-1 sm:hidden">
-                            <PaginationItem>
-                              <PaginationButton
-                                size="sm"
-                                className="px-2.5"
-                                onClick={handlePreviousQuakePage}
-                                disabled={quakePage === 1}
-                                aria-disabled={quakePage === 1}
-                                aria-label="Go to previous quake page"
-                              >
-                                Prev
-                              </PaginationButton>
-                            </PaginationItem>
-                            <PaginationItem className="min-w-0 flex-1 text-center">
-                              <span className="text-xs text-muted-foreground">
-                                {quakePage}/{totalQuakePages}
-                              </span>
-                            </PaginationItem>
-                            <PaginationItem>
-                              <PaginationButton
-                                size="sm"
-                                className="px-2.5"
-                                onClick={handleNextQuakePage}
-                                disabled={quakePage === totalQuakePages}
-                                aria-disabled={quakePage === totalQuakePages}
-                                aria-label="Go to next quake page"
-                              >
-                                Next
-                              </PaginationButton>
-                            </PaginationItem>
-                          </PaginationContent>
-                        </Pagination>
-                      ) : null}
-                    </>
-                  )}
-                </div>
-              </>
-            ) : null}
-
-            {activeMobileDrawer === "global" ? (
-              <>
-                <MobileDrawerHeader
-                  icon={Globe2}
-                  iconToneClassName="bg-amber-500"
-                  title="Global Signals"
-                  description="NASA EONET, OpenAQ, and NWS Tsunami alerts"
-                  onClose={closeDrawer}
-                />
-                <div
-                  data-vaul-no-drag
-                  className="min-h-0 flex-1 overflow-y-auto overscroll-contain px-3 pb-3"
-                >
-                  <div className="min-w-0 rounded-md border border-border/60 bg-muted/15 p-2">
-                    {!layerVisibility.eonet &&
-                    !layerVisibility.airQuality &&
-                    !layerVisibility.tsunami ? (
-                      <div className="rounded-md border border-border/60 bg-background/70 px-3 py-6 text-center text-sm text-muted-foreground">
-                        Global sources are hidden in settings.
-                      </div>
-                    ) : (
-                      <GlobalActivity
-                        collapsed={false}
-                        visibility={{
-                          eonet: layerVisibility.eonet,
-                          airQuality: layerVisibility.airQuality,
-                          tsunami: layerVisibility.tsunami,
-                        }}
-                        eonetState={eonetState}
-                        airQualityState={airQualityState}
-                        tsunamiState={tsunamiState}
-                        onEonetSelect={(event) => {
-                          closeDrawer();
-                          onEonetSelect(event);
-                        }}
-                        onAirQualitySelect={(site) => {
-                          closeDrawer();
-                          onAirQualitySelect(site);
-                        }}
-                        onTsunamiSelect={(alert) => {
-                          closeDrawer();
-                          onTsunamiSelect(alert);
-                        }}
-                      />
-                    )}
-                  </div>
-                </div>
-              </>
-            ) : null}
-
-            {activeMobileDrawer === "weather" ? (
-              <>
-                <MobileDrawerHeader
-                  icon={Cloud}
-                  iconToneClassName="bg-sky-500"
-                  title="Weather"
-                  description="Local conditions and hourly outlook"
-                  onClose={closeDrawer}
-                />
-                <div
-                  data-vaul-no-drag
-                  className="min-h-0 flex-1 overflow-y-auto overscroll-contain px-3 pb-3"
-                >
-                  <div className="min-w-0 space-y-3 rounded-md border border-border/60 bg-muted/15 p-2">
-                  <WeatherDock
-                    latitude={resolvedLocation?.[1] ?? null}
-                    longitude={resolvedLocation?.[0] ?? null}
-                    temperatureUnit={appSettings.temperatureUnit}
-                    timeFormat={appSettings.timeFormat}
-                    unstyled
-                  />
-                  <HourlyForecastDock
-                    latitude={resolvedLocation?.[1] ?? null}
-                    longitude={resolvedLocation?.[0] ?? null}
-                    temperatureUnit={appSettings.temperatureUnit}
-                    timeFormat={appSettings.timeFormat}
-                    className="md:hidden"
-                    compact
-                  />
-                  </div>
-                </div>
-              </>
-            ) : null}
-          </div>
-        </DrawerContent>
-      </Drawer>
-
-      <div className="absolute inset-x-0 bottom-0 pointer-events-none z-50 flex flex-col justify-end pb-safe">
-        <MobileBottomNav
-          items={[
-            {
-              icon: Activity,
-              label: "Quakes",
-              active: activeMobileDrawer === "quakes",
-              onClick: () => activateDrawer("quakes"),
-            },
-            {
-              icon: Cloud,
-              label: "Weather",
-              active: activeMobileDrawer === "weather",
-              onClick: () => activateDrawer("weather"),
-            },
-            {
-              icon: Globe2,
-              label: "Global Meteo",
-              active: activeMobileDrawer === "global",
-              onClick: () => activateDrawer("global"),
-            },
-            {
-              icon: Menu,
-              label: "Menu",
-              active: activeMobileDrawer === "menu",
-              onClick: () => activateDrawer("menu"),
-            },
-          ]}
-        />
-      </div>
-    </div>
-  );
-}
-
-function CustomMapControls({
-  setUserLocation,
-  onLocateAnimation,
-  layerVisibility,
-  onLayerVisibilityChange,
-}: {
-  setUserLocation: (l: [number, number]) => void;
-  onLocateAnimation: () => void;
-  layerVisibility: LayerVisibility;
-  onLayerVisibilityChange: React.Dispatch<React.SetStateAction<LayerVisibility>>;
-}) {
-  const { map } = useMap();
-  const { resolvedTheme, setTheme } = useTheme();
-  const isMobile = useIsMobile();
-  const [waitingForLocation, setWaitingForLocation] = React.useState(false);
-  const [locationFeedbackMessage, setLocationFeedbackMessage] = React.useState<string | null>(null);
-  const compassRef = React.useRef<SVGSVGElement>(null);
-
-  React.useEffect(() => {
-    if (!map) return;
-
-    const updateRotation = () => {
-      if (!compassRef.current) return;
-      const bearing = map.getBearing();
-      const pitch = map.getPitch();
-      compassRef.current.style.transform = `rotateX(${pitch}deg) rotateZ(${-bearing}deg)`;
-    };
-
-    map.on("rotate", updateRotation);
-    map.on("pitch", updateRotation);
-    updateRotation();
-    return () => {
-      map.off("rotate", updateRotation);
-      map.off("pitch", updateRotation);
-    };
-  }, [map]);
-
-  const ease = (t: number) => 1 - Math.pow(1 - t, 3);
-
-  const animateZoom = (delta: number) => {
-    if (!map) return;
-    map.flyTo({
-      zoom: map.getZoom() + delta,
-      duration: 500,
-      easing: ease,
-      curve: 1.3,
-      essential: true,
-    });
-  };
-
-  const handleZoomIn = () => animateZoom(1);
-  const handleZoomOut = () => animateZoom(-1);
-
-  const handleLocate = async () => {
-    if (waitingForLocation) return;
-    if (!map) return;
-
-    setWaitingForLocation(true);
-    setLocationFeedbackMessage(null);
-
-    try {
-      const position = await requestCurrentCoordinates();
-      const coords: [number, number] = [position.longitude, position.latitude];
-      setUserLocation(coords);
-      onLocateAnimation();
-      map.flyTo({
-        center: coords,
-        zoom: 17,
-        duration: 2500,
-        curve: 1.42,
-        speed: 0.6,
-        essential: true,
-        easing: (t) => 1 - Math.pow(1 - t, 3),
-      });
-    } catch (error) {
-      const message = getLocationErrorMessage(error);
-      setLocationFeedbackMessage(message);
-      if (typeof window !== "undefined") {
-        toast.error(message, { duration: 6000 });
-      }
-    } finally {
-      setWaitingForLocation(false);
-    }
-  };
-
-  const handleResetBearing = () =>
-    map?.easeTo({ bearing: 0, pitch: 0, duration: 900, easing: ease, essential: true });
-  const handleFullscreen = () => {
-    const container = map?.getContainer();
-    if (!container) return;
-    if (document.fullscreenElement) document.exitFullscreen();
-    else container.requestFullscreen();
-  };
-
-  const toggleTheme = () => {
-    const current = resolvedTheme ?? "light";
-    setTheme(current === "dark" ? "light" : "dark");
-  };
-
-  const handleToggleLayer = (key: keyof LayerVisibility) => {
-    onLayerVisibilityChange((prev) => ({
-      ...prev,
-      [key]: !prev[key],
-    }));
-  };
-
-  return (
-    <TooltipProvider delayDuration={0}>
-      <div className="flex flex-col gap-2 items-end">
-        <ControlGroup>
-          <DropdownMenu>
-            <Tooltip>
-              <TooltipTrigger asChild>
-                <DropdownMenuTrigger asChild>
-                  <ControlButton label="Filter layers">
-                    <SlidersHorizontal className="size-4" />
-                  </ControlButton>
-                </DropdownMenuTrigger>
-              </TooltipTrigger>
-              <TooltipContent side="left" sideOffset={8}>Filters</TooltipContent>
-            </Tooltip>
-            <DropdownMenuContent align="end" className="w-44">
-              <DropdownMenuLabel>Map Layers</DropdownMenuLabel>
-              <DropdownMenuSeparator />
-              <DropdownMenuCheckboxItem
-                checked={layerVisibility.earthquakes}
-                onCheckedChange={() => handleToggleLayer("earthquakes")}
-              >
-                Earthquakes
-              </DropdownMenuCheckboxItem>
-              <DropdownMenuCheckboxItem
-                checked={layerVisibility.eonet}
-                onCheckedChange={() => handleToggleLayer("eonet")}
-              >
-                NASA EONET
-              </DropdownMenuCheckboxItem>
-              <DropdownMenuCheckboxItem
-                checked={layerVisibility.airQuality}
-                onCheckedChange={() => handleToggleLayer("airQuality")}
-              >
-                Air Quality
-              </DropdownMenuCheckboxItem>
-              <DropdownMenuCheckboxItem
-                checked={layerVisibility.tsunami}
-                onCheckedChange={() => handleToggleLayer("tsunami")}
-              >
-                Tsunamis
-              </DropdownMenuCheckboxItem>
-            </DropdownMenuContent>
-          </DropdownMenu>
-          <Tooltip>
-            <TooltipTrigger asChild>
-              <ControlButton onClick={toggleTheme} label="Toggle theme">
-                {resolvedTheme === "dark" ? (
-                  <Sun className="size-4" />
-                ) : (
-                  <Moon className="size-4" />
-                )}
-              </ControlButton>
-            </TooltipTrigger>
-            <TooltipContent side="left" sideOffset={8}>Theme</TooltipContent>
-          </Tooltip>
-
-          {!isMobile && (
-            <Tooltip>
-              <TooltipTrigger asChild>
-                <ControlButton onClick={handleFullscreen} label="Fullscreen">
-                  <Maximize className="size-4" />
-                </ControlButton>
-              </TooltipTrigger>
-              <TooltipContent side="left" sideOffset={8}>Fullscreen</TooltipContent>
-            </Tooltip>
-          )}
-        </ControlGroup>
-
-        <ControlGroup>
-          <Tooltip>
-            <TooltipTrigger asChild>
-              <ControlButton onClick={handleResetBearing} label="Reset bearing">
-                <svg
-                  ref={compassRef}
-                  viewBox="0 0 24 24"
-                  className="size-5 transition-transform duration-200"
-                  style={{ transformStyle: "preserve-3d" }}
-                >
-                  <title>Compass</title>
-                  <path d="M12 2L16 12H12V2Z" className="fill-red-500" />
-                  <path d="M12 2L8 12H12V2Z" className="fill-red-300" />
-                  <path
-                    d="M12 22L16 12H12V22Z"
-                    className="fill-muted-foreground/60"
-                  />
-                  <path
-                    d="M12 22L8 12H12V22Z"
-                    className="fill-muted-foreground/30"
-                  />
-                </svg>
-              </ControlButton>
-            </TooltipTrigger>
-            <TooltipContent side="left" sideOffset={8}>Reset North</TooltipContent>
-          </Tooltip>
-
-          <Tooltip>
-            <TooltipTrigger asChild>
-              <ControlButton
-                onPointerDown={() => handleLocate()}
-                onTouchStart={() => handleLocate()}
-                onClick={handleLocate}
-                label="Find my location"
-                disabled={waitingForLocation}
-              >
-                {waitingForLocation ? (
-                  <Loader2 className="size-4 animate-spin" />
-                ) : (
-                  <Locate className="size-4" />
-                )}
-              </ControlButton>
-            </TooltipTrigger>
-            <TooltipContent side="left" sideOffset={8}>
-              {locationFeedbackMessage ?? "Your location"}
-            </TooltipContent>
-          </Tooltip>
-        </ControlGroup>
-
-        <ControlGroup>
-          <Tooltip>
-            <TooltipTrigger asChild>
-              <ControlButton onClick={handleZoomIn} label="Zoom In">
-                <Plus className="size-4" />
-              </ControlButton>
-            </TooltipTrigger>
-            <TooltipContent side="left" sideOffset={8}>Zoom In</TooltipContent>
-          </Tooltip>
-
-          <Tooltip>
-            <TooltipTrigger asChild>
-              <ControlButton onClick={handleZoomOut} label="Zoom Out">
-                <Minus className="size-4" />
-              </ControlButton>
-            </TooltipTrigger>
-            <TooltipContent side="left" sideOffset={8}>Zoom Out</TooltipContent>
-          </Tooltip>
-        </ControlGroup>
-      </div>
-    </TooltipProvider>
-  );
-}
-
